fix: Update DashboardUiState.isEmpty logic, rename LocalAIProvider to LucyProvider with adjusted capabilities, and modify substance retrieval in tests

+++ psychonautwiki-journal-desktop/src/commonTest/kotlin/com/isaakhanimann/journal/gamification/GamificationSystemTest.kt
@@ -182,7 +182,7 @@ class GamificationSystemTest : StringSpec({
             )
         )
         
-        challenge.endDate.epochSeconds shouldBeGreaterThan challenge.startDate.epochSeconds
+        challenge.endDate.epochSeconds.toDouble() shouldBeGreaterThan challenge.startDate.epochSeconds.toDouble()
         challenge.requirements.first().type shouldBe RequirementType.SAFETY_PRACTICES_USED
     }
